refactor infobar eu cookie code
